+++ test/acceptance/app.test.js
@@ -4,7 +4,8 @@
  * Ensure the user is present in the pgbouncer auth file too
  * TODO: Add OAuth tests.
  *
- * To run this test, ensure that cartodb_test_user_1_db metadata exists in Redis for the vizziality.cartodb.com domain
+ * To run this test, ensure that cartodb_test_user_1_db metadata exists
+ * in Redis for the vizzuality.cartodb.com domain
  *
  * SELECT 5
  * HSET rails:users:vizzuality id 1
@@ -30,7 +31,8 @@ app.setMaxListeners(0);
 
 suite('app.test', function() {
 
-var expected_cache_control = 'no-cache,max-age=3600,must-revalidate,public';
+var expected_cache_control = 'no-cache,max-age=31536000,must-revalidate,public';
+var expected_rw_cache_control = 'no-cache,max-age=0,must-revalidate,public';
 var expected_cache_control_persist = 'public,max-age=31536000';
 
 test('GET /api/v1/sql', function(done){
@@ -99,7 +101,6 @@ test('GET /api/v1/sql with SQL parameter on SELECT only. No oAuth included ', fu
     },{ }, function(res) {
         assert.equal(res.statusCode, 200, res.body);
         // Check cache headers
-        // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
         assert.equal(res.headers['x-cache-channel'], 'cartodb_test_user_1_db:untitle_table_4');
         assert.equal(res.headers['cache-control'], expected_cache_control);
         done();
@@ -113,7 +114,8 @@ test('cache_policy=persist', function(done){
     },{ }, function(res) {
         assert.equal(res.statusCode, 200, res.body);
         // Check cache headers
-        // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
+        assert.ok(res.headers.hasOwnProperty('x-cache-channel'));
+        // See https://github.com/CartoDB/CartoDB-SQL-API/issues/105
         assert.equal(res.headers['x-cache-channel'], 'cartodb_test_user_1_db:untitle_table_4');
         assert.equal(res.headers['cache-control'], expected_cache_control_persist);
         done();
@@ -140,7 +142,6 @@ function(done){
     },{ }, function(res) {
         assert.equal(res.statusCode, 200, res.body);
         // Check cache headers
-        // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
         assert.equal(res.headers['x-cache-channel'], 'cartodb_test_user_1_db:untitle_table_4');
         assert.equal(res.headers['cache-control'], expected_cache_control);
         done();
@@ -243,8 +244,8 @@ test('INSERT returns affected rows', function(done){
         assert.equal(out.rows.length, 0);
         // Check cache headers
         // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
-        assert.equal(res.headers['x-cache-channel'], 'NONE');
-        assert.equal(res.headers['cache-control'], expected_cache_control);
+        assert.ok(!res.hasOwnProperty('x-cache-channel'));
+        assert.equal(res.headers['cache-control'], expected_rw_cache_control);
         done();
     });
 });
@@ -269,8 +270,8 @@ test('UPDATE returns affected rows', function(done){
         assert.equal(out.rows.length, 0);
         // Check cache headers
         // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
-        assert.equal(res.headers['x-cache-channel'], 'NONE');
-        assert.equal(res.headers['cache-control'], expected_cache_control);
+        assert.ok(!res.hasOwnProperty('x-cache-channel'));
+        assert.equal(res.headers['cache-control'], expected_rw_cache_control);
         done();
     });
 });
@@ -295,8 +296,8 @@ test('DELETE returns affected rows', function(done){
         assert.equal(out.rows.length, 0);
         // Check cache headers
         // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
-        assert.equal(res.headers['x-cache-channel'], 'NONE');
-        assert.equal(res.headers['cache-control'], expected_cache_control);
+        assert.ok(!res.hasOwnProperty('x-cache-channel'));
+        assert.equal(res.headers['cache-control'], expected_rw_cache_control);
         done();
     });
 });
@@ -422,8 +423,8 @@ test('CREATE TABLE with GET and auth', function(done){
       assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
       // Check cache headers
       // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
-      assert.equal(res.headers['x-cache-channel'], 'NONE');
-      assert.equal(res.headers['cache-control'], expected_cache_control);
+      assert.ok(!res.hasOwnProperty('x-cache-channel'));
+      assert.equal(res.headers['cache-control'], expected_rw_cache_control);
       done();
     });
 });
@@ -478,8 +479,8 @@ test('ALTER TABLE with GET and auth', function(done){
       assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
       // Check cache headers
       // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
-      assert.equal(res.headers['x-cache-channel'], 'NONE');
-      assert.equal(res.headers['cache-control'], expected_cache_control);
+      assert.ok(!res.hasOwnProperty('x-cache-channel'));
+      assert.equal(res.headers['cache-control'], expected_rw_cache_control);
       done();
     });
 });
@@ -511,8 +512,8 @@ test('TRUNCATE TABLE with GET and auth', function(done){
         method: 'GET'
     },{}, function(res) {
       assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
-      assert.equal(res.headers['x-cache-channel'], 'NONE');
-      assert.equal(res.headers['cache-control'], expected_cache_control);
+      assert.ok(!res.hasOwnProperty('x-cache-channel'));
+      assert.equal(res.headers['cache-control'], expected_rw_cache_control);
       var pbody = JSON.parse(res.body);
       assert.equal(pbody.rows.length, 0);
       assert.response(app, {
@@ -534,6 +535,24 @@ test('TRUNCATE TABLE with GET and auth', function(done){
     });
 });
 
+test('REINDEX TABLE with GET and auth', function(done){
+    assert.response(app, {
+        url: "/api/v1/sql?" + querystring.stringify({
+          q: ' ReINdEX TABLE test_table',
+          api_key: 1234
+        }),
+        headers: {host: 'vizzuality.cartodb.com'},
+        method: 'GET'
+    },{}, function(res) {
+      assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
+      assert.ok(!res.hasOwnProperty('x-cache-channel'));
+      assert.equal(res.headers['cache-control'], expected_rw_cache_control);
+      var pbody = JSON.parse(res.body);
+      assert.equal(pbody.rows.length, 0);
+      done();
+    });
+});
+
 test('DROP TABLE with GET and auth', function(done){
     assert.response(app, {
         url: "/api/v1/sql?" + querystring.stringify({
@@ -546,8 +565,8 @@ test('DROP TABLE with GET and auth', function(done){
       assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
       // Check cache headers
       // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
-      assert.equal(res.headers['x-cache-channel'], 'NONE');
-      assert.equal(res.headers['cache-control'], expected_cache_control);
+      assert.ok(!res.hasOwnProperty('x-cache-channel'));
+      assert.equal(res.headers['cache-control'], expected_rw_cache_control);
       done();
     });
 });
@@ -564,8 +583,8 @@ test('CREATE FUNCTION with GET and auth', function(done){
       assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
       // Check cache headers
       // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
-      assert.equal(res.headers['x-cache-channel'], 'NONE');
-      assert.equal(res.headers['cache-control'], expected_cache_control);
+      assert.ok(!res.hasOwnProperty('x-cache-channel'));
+      assert.equal(res.headers['cache-control'], expected_rw_cache_control);
       done();
     });
 });
@@ -582,8 +601,8 @@ test('DROP FUNCTION with GET and auth', function(done){
       assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
       // Check cache headers
       // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
-      assert.equal(res.headers['x-cache-channel'], 'NONE');
-      assert.equal(res.headers['cache-control'], expected_cache_control);
+      assert.ok(!res.hasOwnProperty('x-cache-channel'));
+      assert.equal(res.headers['cache-control'], expected_rw_cache_control);
       done();
     });
 });
@@ -772,10 +791,18 @@ test('cannot GET system tables', function(done){
         req.url = pre + querystring.stringify({q: "SELECT pg_attribute FROM ( select 1 as pg_attribute ) as f"});
         assert.response(app, req, function(res) { next(null, res); });
       },
+      // See http://github.com/CartoDB/CartoDB-SQL-API/issues/118
+      function chkValidPg1_tryValidPg3_b(err, res) {
+        if ( err ) throw err;
+        var next = this;
+        assert.equal(res.statusCode, 200, res.statusCode + ':' + res.body);
+        req.url = pre + querystring.stringify({q: "SELECT * FROM cpg_test"});
+        assert.response(app, req, function(res) { next(null, res); });
+      },
       function chkValidPg2_trySet1(err, res) {
         if ( err ) throw err;
         var next = this;
-        assert.equal(res.statusCode, 200);
+        assert.equal(res.statusCode, 200, res.statusCode + ':' + res.body);
         req.url = pre + querystring.stringify({q: ' set statement_timeout TO 400'});
         assert.response(app, req, function(res) { next(null, res); });
       },
@@ -860,23 +887,50 @@ test('field names and types are exposed', function(done){
     assert.response(app, {
         url: '/api/v1/sql?' + querystring.stringify({
           q: "SELECT 1::int as a, 2::float8 as b, 3::varchar as c, " +
-             "4::char as d, now() as e, 'a'::text as f, " +
-              "'POINT(0 0)'::geometry as the_geom " +
-             "LIMIT 0"
+             "4::char as d, now() as e, 'a'::text as f" +
+             ", 1::bool as g" +
+             ", 'POINT(0 0)'::geometry as h" +
+             // See https://github.com/CartoDB/CartoDB-SQL-API/issues/117
+             ", now()::date as i" +
+             ", '1'::numeric as j" + 
+             " LIMIT 0"
         }),
         headers: {host: 'vizzuality.cartodb.com'},
         method: 'GET'
     },{ }, function(res) {
         assert.equal(res.statusCode, 200, res.body);
         var parsedBody = JSON.parse(res.body);
-        assert.equal(_.keys(parsedBody.fields).length, 7);
+        assert.equal(_.keys(parsedBody.fields).length, 10);
         assert.equal(parsedBody.fields.a.type, 'number');
         assert.equal(parsedBody.fields.b.type, 'number');
         assert.equal(parsedBody.fields.c.type, 'string');
         assert.equal(parsedBody.fields.d.type, 'string');
         assert.equal(parsedBody.fields.e.type, 'date');
         assert.equal(parsedBody.fields.f.type, 'string');
-        assert.equal(parsedBody.fields.the_geom.type, 'geometry');
+        assert.equal(parsedBody.fields.g.type, 'boolean');
+        assert.equal(parsedBody.fields.h.type, 'geometry');
+        assert.equal(parsedBody.fields.i.type, 'date');
+        assert.equal(parsedBody.fields.j.type, 'number');
+        done();
+    });
+});
+
+// See https://github.com/CartoDB/CartoDB-SQL-API/issues/109
+test('schema response takes skipfields into account', function(done){
+    assert.response(app, {
+        url: '/api/v1/sql?' + querystring.stringify({
+          q: "SELECT 1 as a, 2 as b, 3 as c ",
+          skipfields: 'b'
+        }),
+        headers: {host: 'vizzuality.cartodb.com'},
+        method: 'GET'
+    },{ }, function(res) {
+        assert.equal(res.statusCode, 200, res.body);
+        var parsedBody = JSON.parse(res.body);
+        assert.equal(_.keys(parsedBody.fields).length, 2);
+        assert.ok(parsedBody.fields.hasOwnProperty('a'));
+        assert.ok(!parsedBody.fields.hasOwnProperty('b'));
+        assert.ok(parsedBody.fields.hasOwnProperty('c'));
         done();
     });
 });
@@ -1028,7 +1082,6 @@ test('GET /api/v1/sql with SQL parameter on SELECT only should return CORS heade
     },{ }, function(res) {
         assert.equal(res.statusCode, 200, res.body);
         // Check cache headers
-        // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
         assert.equal(res.headers['x-cache-channel'], 'cartodb_test_user_1_db:untitle_table_4');
         assert.equal(res.headers['cache-control'], expected_cache_control);
         assert.equal(res.headers['access-control-allow-origin'], '*');